moved view interfaces into their own folder
