--- Enshrouded Server Manager/Models/ViewModels/AdminPanelModel.cs
@@ -1,10 +1,0 @@
-﻿namespace Enshrouded_Server_Manager.Models;
-public class AdminPanelModel
-{
-
-}
-
-public interface IAdminPanelModel
-{
-
-}
