+++ Enshrouded Server Manager/Presenters/ProfileSelectorPresenter.cs
@@ -2,7 +2,7 @@
 using Enshrouded_Server_Manager.Helpers;
 using Enshrouded_Server_Manager.Models;
 using Enshrouded_Server_Manager.Services.Interfaces;
-using Enshrouded_Server_Manager.Views;
+using Enshrouded_Server_Manager.Views.Interfaces;
 
 namespace Enshrouded_Server_Manager.Presenters;
 public class ProfileSelectorPresenter
