+++ Enshrouded Server Manager/Views/Interfaces/IAdminPanelView.cs
@@ -1,6 +1,6 @@
 ﻿using Enshrouded_Server_Manager.Commands;
 
-namespace Enshrouded_Server_Manager.Views;
+namespace Enshrouded_Server_Manager.Views.Interfaces;
 public interface IAdminPanelView
 {
     void SetCommands(IAdminPanelCommand[] commands);
